Move `example/src/App.tsx` to `example/App.tsx`
